+++ example/CirclePlayer.jsx
@@ -1,15 +1,10 @@
 import React, { Component, PropTypes } from 'react'
 import CircleProgress from './CircleProgress'
-import { withMediaPlayer, withMediaProps, controls } from '../src/react-media-player'
+import { Media, Player } from '../src/react-media-player'
 
 class CircleMediaPlayer extends Component {
   componentDidMount() {
     this._circle = new CircleProgress(this._svg)
-  }
-
-  componentDidUpdate() {
-    const { currentTime, duration } = this.props.media
-    this._circle.setProgress(currentTime / duration * 100)
   }
 
   renderPlay() {
@@ -30,25 +25,33 @@ class CircleMediaPlayer extends Component {
     )
   }
 
-  render() {
-    const { Player, media } = this.props
-    const { playPause, isPlaying } = media
+  _handleTimeUpdate = ({ currentTime, duration }) => {
+    this._circle.setProgress(currentTime / duration * 100)
+  }
 
+  render() {
     return (
-      <button className="circle-media-player" onClick={() => playPause()}>
-        { Player }
-        <svg width="32px" height="32px" viewBox="0 0 32 32">
-          <circle cx="16" cy="16" r="14.5" className="circle-media-player__background" />
-          <circle ref={c => this._svg = c} cx="16" cy="16" r="14.5" className="circle-media-player__foreground" />
-          {
-            isPlaying
-              ? this.renderPause()
-              : this.renderPlay()
-          }
-        </svg>
-      </button>
+      <Media
+        src={this.props.src}
+        vendor="audio"
+        onTimeUpdate={this._handleTimeUpdate}
+      >
+        {({ isPlaying, playPause }) =>
+          <button className="circle-media-player" onClick={() => playPause()}>
+            <Player/>
+            <svg width="32px" height="32px" viewBox="0 0 32 32">
+              <circle cx="16" cy="16" r="14.5" className="circle-media-player__background" />
+              <circle ref={c => this._svg = c} cx="16" cy="16" r="14.5" className="circle-media-player__foreground" />
+              { isPlaying
+                  ? this.renderPause()
+                  : this.renderPlay()
+              }
+            </svg>
+          </button>
+        }
+      </Media>
     )
   }
 }
 
-export default withMediaPlayer(withMediaProps(CircleMediaPlayer), 'audio')
+export default CircleMediaPlayer