+++ src/decorators/with-media-props.jsx
@@ -4,7 +4,7 @@ import contextTypes from '../context-types'
 export default function withMediaProps(MediaComponent) {
   return class extends Component {
     static displayName = 'withMediaProps'
-    
+
     static contextTypes = contextTypes
 
     render() {