+++ src/react-media-player.js
@@ -1,5 +1,5 @@
 export Media from './Media'
-export withMediaPlayer from './decorators/with-media-player'
+export Player from './Player'
 export withMediaProps from './decorators/with-media-props'
 export withKeyboardControls from './decorators/with-keyboard-controls'
 export * as controls from './controls/exports'